Parameterize arrays based on count and changed external integer components of the api to be Ints

+++ Sources/LLVMSwift/LLVMModule.swift
@@ -27,6 +27,6 @@ extension LLVMModule {
         name: String = "",
         addressSpace: LLVMAddressSpace = 0
     ) -> LLVMPointer<Value> {
-        LLVMPointer(rawValue: LLVMAddGlobalInAddressSpace(rawModule, type.rawType, name, addressSpace))
+        LLVMPointer(rawValue: LLVMAddGlobalInAddressSpace(rawModule, type.rawType, name, UInt32(addressSpace)))
     }
 }
